modify: rename package from participant to game, and move SentenceCard to game package.
