+++ src/jp/co/ysinc/mojichange/domain/entity/game/Juror.java
@@ -1,4 +1,4 @@
-package jp.co.ysinc.mojichange.domain.entity.participant;
+package jp.co.ysinc.mojichange.domain.entity.game;
 
 /**
  * Created by Fumiya on 2016/05/24.
