+++ src/jp/co/ysinc/mojichange/domain/entity/game/ScoreManager.java
@@ -1,4 +1,4 @@
-package jp.co.ysinc.mojichange.domain.entity.participant;
+package jp.co.ysinc.mojichange.domain.entity.game;
 
 /**
  * Created by Fumiya on 2016/04/02.
