+++ src/jp/co/ysinc/mojichange/domain/repository/PlayerRepository.java
@@ -1,6 +1,6 @@
 package jp.co.ysinc.mojichange.domain.repository;
 
-import jp.co.ysinc.mojichange.domain.entity.participant.Player;
+import jp.co.ysinc.mojichange.domain.entity.game.Player;
 
 /**
  * Created by Fumiya on 2016/04/02.
