--- src/jp/co/ysinc/mojichange/infrastructure/PlayerRepositoryImpl.java
@@ -1,14 +1,0 @@
-package jp.co.ysinc.mojichange.infrastructure;
-
-import jp.co.ysinc.mojichange.domain.entity.participant.Player;
-import jp.co.ysinc.mojichange.domain.repository.PlayerRepository;
-
-/**
- * Created by Fumiya on 2016/04/02.
- */
-public class PlayerRepositoryImpl implements PlayerRepository {
-    @Override
-    public void save(Player player) {
-
-    }
-}
